% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperoverlap-class.R
\name{hyperoverlap-class}
\alias{hyperoverlap-class}
\title{Storage class for the description of hyperoverlaps}
\description{
Storage class for the description of hyperoverlaps
}
\section{Slots}{

\describe{
\item{\code{entity1}}{A length-one character vector}

\item{\code{entity2}}{A length-one character vector}

\item{\code{dimensions}}{A length n character vector containing the variables used to define the space}

\item{\code{occurrences}}{A matrix containing the labelled input data}

\item{\code{shape}}{shape of the decision boundary; either "linear" or "curvilinear"}

\item{\code{polynomial.order}}{a length-one numeric vector showing the polynomial order of the most accurate kernel function. "0" if linear kernel.}

\item{\code{result}}{a length-one character vector, either "overlap" or "non-overlap"}

\item{\code{accuracy}}{a 2x2 table with the true (y) and predicted (pred) labels}

\item{\code{number.of.points.misclassified}}{a length-one numeric vector}

\item{\code{decision.boundary}}{list of 2; used for plotting decision boundary}
}}

