% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperoverlap-package.R
\docType{package}
\name{hyperoverlap-package}
\alias{hyperoverlap-package}
\title{Hyperoverlap: Detection and Visualisation of Overlap in n-Dimensional Space}
\description{
Uses support vector machines to identify a perfectly separating hyperplane (linear or curvilinear) between two entities in high-dimensional space. If this plane exists, the entities do not overlap. Applications include overlap detection in morphological, resource or environmental dimensions. More details can be found in: Brown et al. (2020) \doi{10.1111/2041-210X.13363} .

Uses support vector machines to identify a perfectly separating hyperplane (linear or curvilinear) between two entities in high-dimensional space. If this plane exists, the entities do not overlap. Applications include overlap detection in morphological, resource or environmental dimensions.
}
\details{
More details available in Brown et al. (2020) <doi:10.1111/2041-210X.13363> and vignette.
}
\author{
\strong{Maintainer}: Matilda Brown \email{m.brown2@kew.org} (\href{https://orcid.org/0000-0003-2536-8365}{ORCID})

Authors:
\itemize{
  \item Greg Jordan
}


Matilda Brown \email{matilda.brown@utas.edu.au}
}
\keyword{internal}
