% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperoverlap.r
\docType{package}
\name{hyperoverlap}
\alias{hyperoverlap}
\title{Hyperoverlap: Detection and Visualisation of Overlap in n-Dimensional Space}
\description{
Uses support vector machines to identify a perfectly separating hyperplane (linear or curvilinear) between two entities in high-dimensional space. If this plane exists, the entities do not overlap. Applications include overlap detection in morphological, resource or environmental dimensions.
}
\details{
More details available in our forthcoming paper and vignette.
}
\author{
Matilda Brown \email{matilda.brown@utas.edu.au}
}
