% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperoverlap_plot.r
\name{hyperoverlap_plot}
\alias{hyperoverlap_plot}
\title{Overlap plotting for low-dimensional spaces}
\usage{
hyperoverlap_plot(x)
}
\arguments{
\item{x}{An \code{\link{hyperoverlap-class}} object.}
}
\description{
Plot the optimal separating hyperplane found by hyperoverlap_detect() in 3D .
}
\examples{
data = iris[which(iris$Species!=("versicolor")),]
x = hyperoverlap_detect(data[,1:3],data$Species, kernel="linear")
hyperoverlap_plot(x)

}
\seealso{
\code{\link{hyperoverlap_detect}} , \code{\link{hyperoverlap_lda}}
}
