% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperoverlap_plot.r
\name{hyperoverlap_plot}
\alias{hyperoverlap_plot}
\title{Overlap plotting}
\usage{
hyperoverlap_plot(x)
}
\arguments{
\item{x}{an \code{\link{hyperoverlap-class}} object}
}
\value{
plot of data and decision boundary
}
\description{
Plot the optimal separating hyperplane found by hyperoverlap_detect(). Only used for data of three dimensions or fewer.
}
\examples{
\dontrun{
hyperoverlap_plot(hyperoverlap_iris)
}
}
\seealso{
\code{\link{hyperoverlap_detect}}
}
