% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperoverlap_set.r
\name{hyperoverlap_set}
\alias{hyperoverlap_set}
\title{Pairwise overlap detection in n-dimensional space of multiple entities using support vector machines (SVMs)}
\usage{
hyperoverlap_set(x,y,kernel="polynomial",kernel.degree=3, cost=50,
stoppage.threshold=0.2, write.to.file=TRUE,
path=paste0("hyperoverlap_",Sys.time(),"/"))
}
\arguments{
\item{x}{A matrix or data.frame containing the ecological data}

\item{y}{A vector of labels for each observation}

\item{kernel}{A length-one character vector, either "linear" or "polynomial"}

\item{kernel.degree}{For kernel="polynomial", specifies the order of polynomial kernel function. Smaller orders allow less complex boundary shapes. Values greater than 5 are not recommended for this parameter.}

\item{cost}{Specifies the SVM margin 'hardness'. Default value is 50, but can be increased for improved accuracy, although this increases runtimes and memory usage.}

\item{stoppage.threshold}{Default = 0.2. If the number of points misclassified using a linear hyperplane exceeds this proportion of the number of observations, non-linear separation is not attempted. Can be made 1 to force non-linear separation.}

\item{write.to.file}{TRUE or FALSE, specifies whether each Hyperoverlap object should be saved as a .rds file}

\item{path}{where to write .rfds files to. Defaults to a folder in the working directory titled by Sys.time()}
}
\value{
A long-form matrix with the following columns: entity1, entity2, shape,polynomial.order (if kernel=polynomial), result, number.of.points.misclassified. If specified, individual \code{Hyperoverlap-class} objects are written to file
}
\description{
This function is a wrapper for \code{\link{hyperoverlap_detect}}
}
\examples{
\dontrun{
data(iris)
hyperoverlap.iris.set = hyperoverlap_set(iris[1:3],iris$Species, kernel="linear")
}
}
